moved Manager from people to groups (flask command working)
