+++ api/src/people/test_people.py
@@ -6,7 +6,7 @@ from faker import Faker
 from flask import url_for
 from werkzeug.security import check_password_hash
 
-from .models import Person, PersonSchema, RoleSchema, Role, Manager, ManagerSchema
+from .models import Person, PersonSchema, RoleSchema, Role
 from ..i18n.models import I18NKey, i18n_create, I18NLocale
 from ..images.create_image_data import create_test_images, create_images_people
 from ..images.models import Image, ImagePerson
@@ -1125,267 +1125,6 @@ def test_remove_role_from_account(auth_client):
     # THEN you get a 'That account does not have that role' 404 error
     assert resp.status_code == 404
 
-
-# ---- Manager
-
-@pytest.mark.smoke
-def test_create_manager(auth_client):
-    # GIVEN an empty database
-    person_count = random.randint(10, 20)
-    manager_count = random.randint(5, person_count)
-
-    # WHEN we create a random number of new managers and managers in the database
-    create_multiple_people(auth_client.sqla, person_count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-
-    for i in range(manager_count):
-        resp = auth_client.post(url_for('people.create_manager'),
-                                json=manager_object_factory(auth_client.sqla, 'first level'))
-        assert resp.status_code == 201
-
-    # THEN we end up with the proper number of managers in the database
-    assert auth_client.sqla.query(Manager).count() == manager_count
-
-
-@pytest.mark.smoke
-def test_create_manager_invalid(auth_client):
-    # GIVEN a set of people and accounts
-    count = random.randint(3, 6)
-    create_multiple_people(auth_client.sqla, count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-
-    persons = auth_client.sqla.query(Person).all()
-
-    # GIVEN managers with bad data for each account
-    for person in persons:
-        new_manager = manager_object_factory(auth_client.sqla, fake.sentences(nb=1)[0])
-        new_manager[fake.word()] = fake.word()
-
-        # WHEN a request is made to make a manager with bad data
-        resp = auth_client.post(url_for('people.create_manager'), json=new_manager)
-
-        # THEN expect request to be unprocessable
-        assert resp.status_code == 422
-
-    # THEN expect no managers to be created
-    assert auth_client.sqla.query(Manager).count() == 0
-
-
-@pytest.mark.smoke
-def test_read_all_managers(auth_client):
-    # GIVEN a DB with a collection of managers.
-    person_count = random.randint(10, 20)
-    manager_count = random.randint(5, person_count)
-    create_multiple_people(auth_client.sqla, person_count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, manager_count)
-    # WHEN we request all managers from the server
-    resp = auth_client.get(url_for('people.read_all_managers', locale='en-US'))
-    # THEN the count matches the number of entries in the database
-    assert resp.status_code == 200
-    assert len(resp.json) == manager_count
-
-
-@pytest.mark.smoke
-def test_read_one_manager(auth_client):
-    # GIVEN a DB with a collection of managers.
-    person_count = random.randint(10, 20)
-    manager_count = random.randint(5, person_count)
-    create_multiple_people(auth_client.sqla, person_count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, manager_count)
-
-    # WHEN we ask for them all
-    managers = auth_client.sqla.query(Manager).all()
-
-    # THEN we expect the same number
-    assert len(managers) == manager_count
-
-    # WHEN we request each of them from the server
-    for manager in managers:
-        resp = auth_client.get(url_for('people.read_one_manager', manager_id=manager.id, locale='en-US'))
-        # THEN we find a matching manager
-        assert resp.status_code == 200
-        assert resp.json['person_id'] == manager.person_id
-        assert resp.json['manager_id'] == manager.manager_id
-        assert resp.json['description_i18n'] == manager.description_i18n
-
-
-@pytest.mark.smoke
-def test_read_one_manager_invalid(auth_client):
-    # GIVEN an empty database
-
-    # WHEN a request is made to read a manager that does not exist
-    resp = auth_client.get(url_for('people.read_one_manager', manager_id=random.randint(1, 8)))
-
-    # THEN expect requested manager not to be found
-    assert resp.status_code == 404
-
-
-@pytest.mark.smoke
-def test_update_manager(auth_client):
-    # GIVEN a DB with a collection of managers.
-    person_count = random.randint(10, 20)
-    manager_count = random.randint(5, person_count)
-    create_multiple_people(auth_client.sqla, person_count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, manager_count)
-
-    managers = auth_client.sqla.query(Manager).all()
-    persons = auth_client.sqla.query(Person).all()
-
-    update_manager = random.choice(managers)
-
-    new_person_id = update_manager.person_id
-    while new_person_id == update_manager.person_id:
-        new_person_id = random.choice(persons).id
-
-    new_manager_id = update_manager.manager_id
-    while new_manager_id == update_manager.manager_id or new_manager_id == update_manager.id:
-        new_manager_id = random.choice(managers).id
-
-    update_json = {
-        'person_id': new_person_id,
-        'manager_id': new_manager_id,
-        'description_i18n': update_manager.description_i18n
-    }
-
-    # WHEN
-    resp = auth_client.patch(url_for('people.update_manager', manager_id=update_manager.id), json=update_json)
-    # THEN
-    assert resp.status_code == 200
-    assert resp.json['person_id'] == new_person_id
-    assert resp.json['manager_id'] == new_manager_id
-
-
-@pytest.mark.smoke
-def test_update_manager_invalid(auth_client):
-    # GIVEN a set of people, accounts, and managers
-    count = random.randint(3, 6)
-    create_multiple_people(auth_client.sqla, count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, count)
-
-    managers = auth_client.sqla.query(Manager).all()
-
-    # GIVEN bad modification data
-    for manager in managers:
-        mod = {}
-        flips = (flip(), flip(), flip())
-        if flips[0]:
-            mod['person_id'] = None
-        if flips[1]:
-            mod['description_i18n'] = None
-        if flips[2] or not (flips[0] or flips[1]):
-            mod[fake.word()] = fake.word()
-
-        # WHEN a manager is requested to be updated with bad data
-        resp = auth_client.patch(url_for('people.update_manager', manager_id=manager.id), json=mod)
-
-        # THEN expect the request to be unprocessable
-        assert resp.status_code == 422
-
-
-@pytest.mark.smoke
-def test_delete_manager(auth_client):
-    # GIVEN a DB with a collection of managers.
-    person_count = random.randint(10, 20)
-    manager_count = random.randint(5, person_count)
-    create_multiple_people(auth_client.sqla, person_count)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, manager_count)
-
-    managers = auth_client.sqla.query(Manager).all()
-    persons = auth_client.sqla.query(Person).all()
-
-    delete_manager = managers[0]
-    subordinate = managers[1]
-
-    update_json = {
-        'manager_id': delete_manager.id
-    }
-    auth_client.patch(url_for('people.update_manager', manager_id=subordinate.id), json=update_json)
-
-    # WHEN we delete the manager
-    resp = auth_client.delete(url_for('people.delete_manager', manager_id=delete_manager.id))
-
-    # THEN the manager and all references to that manager are deleted
-    assert resp.status_code == 204
-    assert auth_client.sqla.query(Manager).filter_by(id=delete_manager.id).first() == None
-    assert auth_client.sqla.query(Manager).filter_by(id=subordinate.id).first().manager_id == None
-
-
-@pytest.mark.smoke
-def test_delete_manager_no_exist(auth_client):
-    # GIVEN an empty database
-
-    # WHEN a manager is requested to be deleted
-    resp = auth_client.delete(url_for('people.delete_manager', manager_id=random.randint(1, 8)))
-
-    # THEN expect the manager not to be found
-    assert resp.status_code == 404
-
-
-@pytest.mark.smoke
-def test_delete_manager_with_subordinate(auth_client):
-    # GIVEN a set of people, accounts, and managers
-    count = random.randint(3, 6)
-    create_multiple_people(auth_client.sqla, count * 2)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, count * 2)
-
-    # GIVEN half of the managers are subordinates of the others
-    for i in range(1, count + 1):
-        auth_client.sqla.query(Manager).filter(Manager.id == i).update({"manager_id": i + count})
-        auth_client.sqla.commit()
-
-    # WHEN superiors are requested to be deleted
-    for i in range(1, count + 1):
-        resp = auth_client.delete(url_for('people.delete_manager', manager_id=i + count))
-
-        # THEN expect the delete to run OK
-        assert resp.status_code == 204
-
-        # THEN expect that superior was deleted
-        assert auth_client.sqla.query(Manager).count() == count * 2 - i
-
-        # THEN expect relationship with subordinate to be broken
-        assert auth_client.sqla.query(Manager).filter(Manager.id == i).first().manager_id is None
-
-
-@pytest.mark.smoke
-def test_create_manager_with_superior(auth_client):
-    # GIVEN a set of people, accounts, and managers
-    count = random.randint(3, 6)
-    create_multiple_people(auth_client.sqla, count * 2)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, count)
-
-    superiors = auth_client.sqla.query(Manager).all()
-
-    # GIVEN data for subordinate managers
-    for superior in superiors:
-        new_manager = manager_object_factory(auth_client.sqla, fake.sentences(nb=1)[0])
-        new_manager['manager_id'] = superior.id
-
-        # WHEN subordinates are requested to be created
-        resp = auth_client.post(url_for('people.create_manager'), json=new_manager)
-
-        # THEN expect the create to run OK
-        resp.status_code == 201
-
-    # THEN expect the right number of managers to be created
-    assert auth_client.sqla.query(Manager).count() == count * 2
-
-
-@pytest.mark.smoke
-def test_repr_manager(auth_client):
-    # GIVEN a DB with a manager
-    create_multiple_people(auth_client.sqla, 1)
-    #create_multiple_accounts(auth_client.sqla, 1)
-    create_multiple_managers(auth_client.sqla, 1)
-    managers = auth_client.sqla.query(Manager).all()
-    managers[0].__repr__()
 
 
 @pytest.mark.smoke
